+++ Classes/Extension.php
@@ -21,7 +21,5 @@ final class Extension
     public const LANGUAGE_PATH_DATABASE = 'LLL:EXT:' . self::KEY . '/Resources/Private/Language/locallang_db.xlf';
     public const LANGUAGE_PATH_DEFAULT = 'LLL:EXT:' . self::KEY . '/Resources/Private/Language/locallang.xlf';
 
-    public const CACHE_IDENTIFIER = 'tx_' . self::KEY;
-
     public const JSONLD_TEMPLATE = '<script type="application/ld+json" id="ext-schema-jsonld">%s</script>';
 }